Fix panic when too many slabs are batched
While slabs are loaded, they may block on waiting for dependent
neighbours to load, which relies on receiving broadcasted status updates
via a channel. If that channel is full then updates are missed, so now
instead of unconditionally failing, try again if the channel is
full (including the check of "do we need to keep waiting at all").